third_party/blink/web_tests/html/dialog/multiple-centered-dialogs-expected.txt

Test that multiple dialogs are centered properly.
showing top-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing first-middle-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing second-middle-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing bottom-dialog
PASS dialog.offsetTop is expectedTop(dialog)
PASS successfullyParsed is true

TEST COMPLETE
